Move the jdbcRDD methods into SQLContext itself.
