+++ sql/core/src/main/scala/org/apache/spark/sql/jdbc/JDBCRelation.scala
@@ -34,7 +34,7 @@ private[sql] case class JDBCPartition(whereClause: String, idx: Int) extends Par
 /**
  * Instructions on how to partition the table among workers.
  */
-private[jdbc] case class JDBCPartitioningInfo(
+private[sql] case class JDBCPartitioningInfo(
     column: String,
     lowerBound: Long,
     upperBound: Long,
